+++ ts/test/main.test.ts
@@ -156,8 +156,7 @@ describe("composite-map", () => {
         expect((map as any).copiedSet).to.equal(undefined);
         const mapCopy = new CompositeMap(map, { copy: "on-write" });
 
-        // expect((map as any).copiedSet).to.be.a("Set");
-        expect((mapCopy as any).copiedSet).to.be.a("Set");
+        expect((mapCopy as any).copiedSet).to.be.a("WeakSet");
         expect(mapCopy.get(["a", "b"])).to.equal("test");
         expect(mapCopy.get([])).to.equal(map.get([]), "The unmodified copy should use the same root map object");
 
@@ -205,6 +204,39 @@ describe("composite-map", () => {
         expect(mapCopy.has(["b", "a"])).to.equal(true, "Changes to the original map should not affect the copy");
     });
 
+    it('constructor(map, { copy: "on-write" }), .prototype.delete(zero_length_key)', () => {
+        const map = new CompositeMap<string, string>();
+        // tslint:disable-next-line:no-unused-expression
+        new CompositeMap(map, { copy: "on-write" });
+
+        expect((map as any).copiedSet).to.be.a("WeakSet");
+        expect(map.delete([])).to.equal(false, "delete when the map has no keyLength");
+        expect((map as any).copiedSet).to.equal(undefined);
+
+        map.set(["a"], "test");
+        // tslint:disable-next-line:no-unused-expression
+        new CompositeMap(map, { copy: "on-write" });
+        expect((map as any).copiedSet).to.be.a("WeakSet");
+        (map.get([]) as Map<string, string>).clear();
+        expect(map.delete([])).to.equal(false, "delete when the map has a keyLength but no records");
+        expect((map as any).copiedSet).to.equal(undefined);
+    });
+
+    it('constructor(map, { copy: "on-write" }), .prototype.clear()', () => {
+        const map = new CompositeMap<string, string>();
+        // Note: Unless a sub-map has multiple entries, pruning will cause the sub-map to remain unchanged since it
+        //   would instead be deleted by its parent.
+        map.set(["a", "a"], "test");
+        const mapCopy = new CompositeMap(map, { copy: "on-write" });
+
+        expect(map.has(["a", "a"])).to.equal(true);
+        expect((map as any).copiedSet).to.be.a("WeakSet");
+        map.clear();
+        expect(map.has(["a", "a"])).to.equal(false);
+        expect((map as any).copiedSet).to.equal(undefined);
+        expect(mapCopy.has(["a", "a"])).to.equal(true);
+    });
+
     it('constructor(map, { copy: "keys" })', () => {
         const map = new CompositeMap<string, string>();
         map.set(["a", "b"], "test").set(["b", "a"], "test2");
@@ -218,20 +250,6 @@ describe("composite-map", () => {
         mapCopy.set(["a", "b"], "test3");
         expect(mapCopy.get(["a", "b"])).to.equal("test3");
         expect(map.get(["a", "b"])).to.equal("test", "The original map should remain unchanged");
-    });
-
-    it('constructor(map, { copy: "reference" })', () => {
-        const map = new CompositeMap<string, string>();
-        map.set(["a", "b"], "test").set(["b", "a"], "test2");
-        const mapCopy = new CompositeMap(map, { copy: "reference" });
-
-        expect((mapCopy as any).copiedSet).to.equal(undefined);
-        expect(mapCopy.get(["a", "b"])).to.equal("test");
-        expect(mapCopy.get([])).to.equal(map.get([]), "The copy should use the same root map object");
-
-        mapCopy.set(["a", "b"], "test3");
-        expect(mapCopy.get(["a", "b"])).to.equal("test3");
-        expect(map.get(["a", "b"])).to.equal("test3", "The original map should also be changed");
     });
 
     it('constructor(map, { copy: "invalid" })', () => {