+++ changelog.md
@@ -3,3 +3,9 @@
 ## 1.0.0 (2019-01-20)
 
 -   initial release
+
+## 1.1.0 (2019-02-03)
+
+-   bugfix: Removed copy "reference" option.
+-   bugfix: Clear method when using copy-on-write should not affect the original data.
+-   bugfix: Copy-on-write object are garbage collected when deleted directly.